Refactored GameVariableNames ad updated collisions
Refactored GameVariableNames :
Now, all the game variables are stored in the GameVariableNames class.
It is easier to change the name of a variable in the game and to call it
=> it removes the risk of typos.
The collisions have been updated to use the new GameVariableNames class.

Updated collisions
Now, a score is displayed for the current player.
It increases when the player kills an alien.
Aliens can't shoot for the moment. But if they shoot, the player will
lose a life.
The life of the player is also displayed.
Refactored the necessary classes to be up to date with modifications.

Fixed the issue on the Integer value for the lives.