refactor(GameMode): :recycle: refactor code for easier game mode implementation
I refactored the game modes in a simple class.
Also created an interface for the game modes.
The GameLauncher now uses the GameMode class to launch the game.
It only depends on the GameMode interface.

WIP : the game isn't working yet. Some important fixes need to be made.
The inputs aren't correctly working for now.