refactor(GameMode): :recycle: heavy refactor
This commit fixes a lot of issues with the game mode system.
I extracted all the game mode logic into abstract classes corresponding
to the number of players for the game mode.
Moreover, I fixed the GameLauncher for displaying the update on players'
life and points.
Removed some unused GameVariableNames.