fix(input): :ambulance: finally fixed all inputs !
:tada: I finally got the right perception to fix this bug !!!
The solution is to put the inputs in the gameLauncher. The binding will
ask the game mode the playerComponent that will be used for the active
game. So, it is very simple as the One player or Two player game modes
can return either the playerComponent1 or the playerComponent2.
This is pure awesomeness ! :tada: